1. Changed create method in generic DAO from int to Serializable for working with different types of key. 2. Added PasswordChangeRequest entity for storing reset password links. 3. Added confirmation email for registration 4. Added thymeleaf spring security integration

+++ src/main/java/com/kverchi/diary/service/impl/BookServiceImpl.java
@@ -34,7 +34,7 @@ public class BookServiceImpl implements BookService {
 	@Override
 	
 	public Book addBook(Book book) {
-		int addedId = bookDao.create(book);
+		int addedId = (Integer)bookDao.create(book);
 		Book addedBook = bookDao.getById(addedId);
 		return addedBook;
 	}
